Extract ByteArray wrapper to avoid exposing QByteArray and QHash
