+++ src/App/CMakeLists.txt
@@ -295,6 +295,7 @@ SET(FreeCADApp_CPP_SRCS
     ApplicationPy.cpp
     AutoTransaction.cpp
     Branding.cpp
+    ByteArray.cpp
     CleanupProcess.cpp
     ColorModel.cpp
     ComplexGeoData.cpp
@@ -328,6 +329,7 @@ SET(FreeCADApp_HPP_SRCS
     ApplicationDirectories.h
     AutoTransaction.h
     Branding.h
+    ByteArray.h
     CleanupProcess.h
     ColorModel.h
     ComplexGeoData.h
