+++ src/App/IndexedName.h
@@ -30,11 +30,9 @@
 #include <string>
 #include <vector>
 
-#include <QByteArray>
-#include <QHash>
-
 #include "FCGlobal.h"
 
+class QByteArray;
 
 namespace Data
 {
@@ -135,12 +133,7 @@ public:
      *
      * @param[in] data The QByteArray to copy the data from.
      */
-    explicit IndexedName(const QByteArray& data)
-        : type("")
-        , index(0)
-    {
-        set(data.constData(), data.size());
-    }
+    explicit IndexedName(const QByteArray& data);
 
     /**
      * @brief Create an indexed name from a string and index.
@@ -402,67 +395,6 @@ protected:
 private:
     const char* type;
     int index;
-};
-
-
-/// A thin wrapper around a QByteArray providing the ability to force a copy of the data at any
-/// time, even if it isn't being written to. The standard assignment operator for this class *does*
-/// make a copy of the data, unlike the standard assignment operator for QByteArray.
-struct ByteArray
-{
-    explicit ByteArray(QByteArray other)
-        : bytes(std::move(other))
-    {}
-
-    ByteArray(const ByteArray& other) = default;
-
-    ByteArray(ByteArray&& other) noexcept
-        : bytes(std::move(other.bytes))
-    {}
-
-    ~ByteArray() = default;
-
-    /// Guarantee that the stored QByteArray does not share its memory with another instance.
-    void ensureUnshared() const
-    {
-        QByteArray copy;
-        copy.append(bytes.constData(), bytes.size());
-        bytes = copy;
-    }
-
-    bool operator==(const ByteArray& other) const
-    {
-        return bytes == other.bytes;
-    }
-
-    ByteArray& operator=(const ByteArray& other)
-    {
-        bytes.clear();
-        bytes.append(other.bytes.constData(), other.bytes.size());
-        return *this;
-    }
-
-    ByteArray& operator=(ByteArray&& other) noexcept
-    {
-        bytes = std::move(other.bytes);
-        return *this;
-    }
-
-    mutable QByteArray bytes;
-};
-
-
-struct ByteArrayHasher
-{
-    std::size_t operator()(const ByteArray& bytes) const
-    {
-        return qHash(bytes.bytes);
-    }
-
-    std::size_t operator()(const QByteArray& bytes) const
-    {
-        return qHash(bytes);
-    }
 };
 
 }  // namespace Data
